DRM: check key statuses initially linked to a MediaKeySession
This fixes an issue in the v3.30.0 - for now only in
pre-release - and v4.0.0-beta.0.

When playing a content whose MediaKeySession(s) were already cached and
which need to have Representation fallback(s) (some Representation are
not decipherable), it was possible to be left in an infinite rebuffering
case.

The reason is very simple and a classic issue: since the removal of RxJS
in the `SessionEventListener` code, the DRM code only alerted of
unusable key-ids after the `keyStatusesChange` `MediaKeySession` event
was triggered. Thus it would not alert of unusable key ids if all
statuses were already known from the beginning - such as in a reused
cached `MediaKeySession` case.

The fix is straightforward and is similar to the logic before
the removal of RxJS: as soon as the `SessionEventListener` is called, we
begin to check the key statuses already there - and not only on the
`onKeyStatusesChange` event.