For dynamic contents, always set a very high duration
I found out that the video element's `duration` property for live
contents was sometimes very close to the maximum position, whereas
it should be set to very high number for such types of contents.

The reason behind this issue was very clear: the `MediaDurationUpdater`
only performed a live-specific logic when the current tracks' duration
are initially not known whereas it should do it everytime the duration
may need to be updated.