Set `Infinity` as duration when playing live contents on the PS5
We recently noticed an issue on the PS5 which causes stuttering playback
on live contents.

After exchanging with Sony, turns out the root cause is that their new
WebKit version has performance troubles in its segment data eviction
logic when the current `MediaSource`'s duration is set at a high value.

For live contents, we set in most cases (there are rare exceptions) the
duration to 2^32, because higher values may cause issues on other
platforms (most notably Tizen). To us, it didn't seem as a very high
value but Sony actually advised us to use positive `Infinity` instead (a
valid IEEE 754 value that can be set in JavaScript for a "Number"), as
it seems to be handled specially by WebKit.

`Infinity` is also the actual value advised by the current HTML living
standard for live contents, but I was still afraid that just changing to
that special value would break other platforms - as we're used to random
bugs when changing those types of properties. Because of this, and despite
the fact that the issue seems rooted in a WebKit version, I chose to
only set the `Infinity` duration on the PS5 for now as this is the only
device where that issue has been seen for now.