Avoid destroying PeriodStream if one is re-created for the same period
While doing some debugging on multi-Period contents, we saw some
situations where at some Period's boundary, no segment would be loaded,
since the `v3.30.0`.

Though we are still in the process of testing it, I relatedly found what
seems like an issue where we could be in some kind of (slow) loop creating
and destroying `PeriodStream`, which is the module allowing to load
segment corresponding to a particular DASH period (or related in other
`transport`s) when we approach from the end of the previous Period.

To try to fix this logic, I added a check to see if the next
`PeriodStream` wanted was linked to the same `Period` than the next
`PeriodStream` which was already created, in which case we won't be
destroying it immediately, just to re-creating it like we did before.

For the source of this behavior, I think that re-checking that the right
`PeriodStream` is created regularly (more specifically: each time the previous
chronological `PeriodStream` signals that is has no segment to load) is not
that bad of an idea, as a security, so I did not change it.