DASH/WASM: actually reject the DASH_WASM.initialize's Promise if it fails
I forgot to add a line on the implementation of the `DASH_WASM.initialize`J
API (which allows to "initialize" the WASM-based DASH parser, i.e. load
and compile its linked WebAssembly file), which led to the Promise returned
by this method still resolved even if initialization fails.

The error is still logged and the internal status of the `DASH_WASM`
feature is still updated to indicate failure, but that's not what the
application is most likely listening to.