Externalize our doc generator
This has no need to be in the RxPlayer's source code and we wanted to
externalize it for quite some time.

The main reasons are that it only has a very peripheral relation to the
RxPlayer, it represents a lot of code, most of its dependencies are only
used by it and that other projects may well also rely on it if they want
to.

I took the occasion of a personal project to ease-up externalization (on
my free time) and now it is published on `npm` with the name
`docgen.ico` - a somewhat dubious play-on-word that I won't explain
here.

I feel a little bad that its original repo's is in my personal GH repos,
but sadly, this has been a long time tradition at Canal+ as creating
public repos under the Canal+ company has been impossible despite
multiple demands. Multiple of our work-related PoCs and side-projects
have generally been written on personal pages like that, though one
has been migrated to Canal+'s gitlab.

In a perfect world, the one relied on by the RxPlayer should be under
the `Canal+` company - and may be registered under another name, and I
may maintain my own fork as a personal project for quick iteration on my
other personal projects - while upstreaming PRs to Canal+ when I do
that.

However even with all this considered, I did not maintain that doc
generator with security, customizability nor even code quality in mind,
it always has been a quick hack to have a simple doc generator which
only rely on very simple markdown files and a few json configuration
files, as alternatives either proposed some custom formats or were
hugely bloated and complicated for our need (especially Docusaurus,
which had kind of both).
If it is really relied on by more projects, there WILL be bugs and
sadness!

Though now that an official repo and npm package exists, I may look at
this project with a different eye in future versions.