DASH_WASM: Update quick-xml (Rust) dependency and consequently update Rust code
This commit updates the dependency in our Rust code, `quick-xml`, which
allows to parse XML documents (which DASH MPD are) in our WASM MPD
parser in a SAX-like manner.

The major change here is that the types for element names and attributes
is now implemented by single-field structs whose field is the reference
to the corresponding name whereas it was directly the reference initially.

Thankfully, there's no need to bring that new struct's name in all our
pattern-matching cases (which would have been less readable), as calling
the `as_ref` method implemented on that struct returns directly the
reference we had before (we can thus pattern match on this instead, and
keep the previous code for all matched patterns).
This is actually what is presented in quick-xml examples so it also seems
like the idiomatic way to do it.

However the major part of that code is a formatting update, done thanks
to the rust-official `rustfmt` tool and code updates based on clippy
(official Rust linter) lints.

This branch was tested with success on various DASH contents.