Update "list" npm script so it may actually execute commands
This is an update of the "list" npm script, which brings two things to
improve its usefulness:
  - if less than 10 numbered inputs are possible, the script won't wait
    for the enter key to be entered for validating your choice.
  - Listed commands are now choosable, and choosing one will execute it.

This is to allow new RxPlayer developpers to more easily familiarize
themselves to the high number of scripts there are in the RxPlayer (to
build, build a demo, build the doc, doing the various kinds of tests -
unit, integration, memory & performance -, doing the various kinds of
code checks - linting, typechecking -, doing some maintenance, updating
the version, updating the served demo and documentation pages and so
on...).

Here the idea is to just use a single script at the beginning, `npm run
list`, which in two steps asks you:
  1. what you want to do (e.g. build a demo, run tests etc.)
  2. list all script variants to do just that and let you directly run
     that script.

Once a developper becomes familiar enough with the various scripts,
he/she may start using them directly instead.