Remove dist folder and its content from git for the v4
This is a proposal for something I wanted to do for a long time but
didn't for legacy reasons.

Today builds, which are parts of the `dist/` folder at the root of this
repository, are stored and its update committed through git, alongside the
rest of the code.

This means that people cloning or even just browsing the repo is going
to be polutted by megabytes and megabytes of uninteresting bundled code
which can already be built on demand through our bundling dependencies
(I hope deterministically? Or at least not different in ways that
matters?).

Removing the dist folder completely would have the following pros, at
least to me:

  - no need to store the bundle through git and its mostly uninteresting
    and large diffs, which may lead to a future gain in storage space,
    though I don't know what the proportions are (chances are they're
    very small, but still).

  - When using git tools or tools relying on files like `.gitignore`
    (e.g. the searching tool ripgrep - which I use daily. I think that
    VSCode also rely on it or at least it could last time I read about
    it - for the lost and sad people that are not on vim/neovim yet).

    Before, using those tools on the whole repo without explicitely
    discarding the `dist` directory would lead to a lot of polution,
    especially in search.

  - It un-clutters the repository. I like the idea of being able to
    understand its structure just by browsing it and one less directory
    is a nice step towards that goal.

Yet it has the following cons I can think of:

  - It removes the previous possibility of just looking at the build on
    github and copy-pasting the bundle's code. Which some did on very
    fast tests in the past.

  - It leads to the necessity of first bundling the rx-player before
    doing things like `npm link`/`yarn link` or other tools/scripts that
    I'm not thinking of right now.