Improve global scope situation with util and lint update
Some years back, we were notified of an issue when the RxPlayer was
imported in a Node.JS environment - which has mostly no chance to bring
any result unless that environment implements MSE, EME and DOM API - but
which was still done due to some logic linked to server-side-rendering.

That issue lead to an error, and was due to the RxPlayer refering to the
`window` (the global scope accessible in JavaScript in a main thread
environment) at the top level (generally done there to check once at
script evaluation if various browser API are present), because `window`
does not exist in Node.JS environments.

What we did at the time is just to check (poorly) if we were in a
Node.js environment, and to not rely on `window` if that was the case at
those various places. However that was error-prone (we could forget that
check) and the only thing preventing mistakes was a script in
`scripts/check_nodejs_import_compatibility.js` which wouldn't be able
to catch all cases.

---

Because this became again a problem when PoCing worker environments, I
profited from this project to work on a better solution.

Now the global scope is directly provided by an util function of our
compat code, which works for NodeJS, JS main thread and JS WebWorker
environments, and its usage is enforced by a linter rule which should
hopefully catch all `window` usage in the code.

I also chose to erase all usage of the term `window` or the shortened
`win` alias in code to better communicate the point that we're
specifically mean the global scope, regardless of how it is referred to
in the current environment.