Implement better messages for request errors
While debugging some issues, we noticed that our network-related error
messages could do with more verbosity. This is what this commit does.

It also adds the HTTP status code in the message when this is the
source, of the network error. Today it is only present as a `status`
property.