[Proposal] "Remove" `manifestUpdateUrl` loadVideo option for the v4
This commit undocument the `manifestUpdateUrl` API, so it is effictevely
not part of the official API anymore, even if its behavior stays for now
(renamed as `__priv_manifestUpdateUrl`).

The `manifestUpdateUrl` was an advanced DASH optimization where it was
possible to update a dynamic MPD through a shorter version of it, thus
much faster.
This was a clever solution for a real issue we had for a long time at
Canal+, but it relied on complex and easy-to-break logic to perform
updates.

Now, we prefer other solutions, which are either standard (e.g. DASH
packaging techniques to reduce the size of the MPD) or purely based on
code optimizations (WebAssembly-based parsers, advanced parsing modes) -
thus not relying on special URL like it was with this.

As It is not needed at Canal+ anymore, it is not standard in any way,
we didn't hear of any other person using this feature and as removing
it does not break playback (it only make it less efficient if the
feature was relied on before), I propose we profit from the v4 still
being in beta to remove it before an official v4.

Though this commit does not really remove the corresponding logic yet,
it only moves it as an undocumented API.
This is because we may want to still keep that logic maintained for
some months in case its need ever seems to make a comeback.