Fix theoretical memory leak on reload in stream PlaybackObserver creation code
When working on the WebWorker branch (#1272), I noticed a (small) memory leak
where a created PlaybackObserver could still be running after a reload
for the same content, whereas it should not.

This removes that leak by ensuring all logic is cleaned-up on each
reload.

However, we do have memory-leak tests checking for leaks in that area
and it didn't find anything, running the corresponding test before and after
the commit is applied seems to result in roughly the same JS heap size.
We may have to look if the test does not do what we want it to do or if the
leak in question is very insignifance/does not apply in our current code.