Fix error in improbable case where content is already ready on load
The code implementing the initial seek (e.g. to play the initial
position of a live content, or to implement the `startAt` option) and
play (e.g. to implement `autoPlay`) could lead to an error in the
for-now improbable (impossible?) scenario where the HTMLMediaElement
already has at least the `HAVE_METADATA` `readyState` before that
function is called, due to a callback using a variable before it was
defined.

I don't know of any current issue that may be triggered by this, but
some future work triggered it, so let's just fix that future issue.