StreamEventsEmitter is now a class
The `StreamEventsEmitter` is the RxPlayer module storing events found in
the Manifest (for now only DASH' `EventStream` elements) and emitting
`streamEvent`/`streamEventSkip` RxPlayer events when they are encountered.

In the WebWorker work (#1272), I had to refacto the
`StreamEventsEmitter` from being a long-lived function to being a class,
to make it possible to be communicated manifest updates less awkwardly
and more idiomatically, through an `onManifestUpdate` method.

However, I liked the result and find it much more easy to understand
than the previous one, which bears the mark of having been previously
an RxJS-based logic.

This refacto the `StreamEventsEmitter` into a class without adding the
supplementary worker-linked logic handlin manifest updates differently.

Basically, it's just now a class emitting events with `start` and `stop`
methods.