RepresentationStream: Also check segments after a Manifest update
The mechanism to check if new segments are available are done on several
events but was not done on a Manifest update.

Even if for complex reasons, there is small advantages to not check for new
segments directly after a Manifest update, it also seems in most cases more
logical and efficient.

I decided to add it.