Centralize debug mode handling
The RxPlayer has an hidden feature of setting automatically debug logs
if a `__RX_PLAYER_DEBUG_MODE__` variable has been set to `true` in the
globak scope. This allows to ease-up debugging work by preventing people
to always having to manually set the log verbosity of the RxPlayer, and
it is used by [RxPaired](https://github.com/peaBerberian/RxPaired), the
RxPlayer-specialized remote debugger we routinely use and develop on.

Because in most scenarios, `__RX_PLAYER_DEBUG_MODE__` is not even
defined as a global variable, just checking it that way will throw an
error in JavaScript, so we have to be careful when relying on it: we
HAVE TO rely on the `typeof` operator first, as it doesn't lead to an
error if the variable given a argument does not exist.

However, TypeScript doesn't seem to have any mean to indicate that a
global variable may or not be globally available, so we just had to
manually remind ourselves to call `typeof` on it before.

To be less error-prone, I chose to isolate debug-mode checking in its
own util and to depend on it everywhere else.